+++ CHANGES.rst
@@ -12,6 +12,9 @@ New Features
     (Dec, elevation, galactic B) respectively.  The `Longitude` class provides
     auto-wrapping capability and `Latitude` performs bounds checking.
 
+  - `astropy.coordinates.Distance` supports conversion to and from distance
+    modulii.
+
 - ``astropy.io.votable``
 
   - The format of the units of a VOTable file can be specified using
@@ -102,7 +105,13 @@ API Changes
 - ``astropy.coordinates``
 
   - The `astropy.coordinates.Angle` class is now a subclass of
-    `astropy.units.Quantity`.
+    `astropy.units.Quantity`. This means it has all of the methods of a
+    `numpy.ndarray`.
+
+
+  - The `astropy.coordinates.Distance` class is now a subclass of
+    `astropy.units.Quantity`. This means it has all of the methods of a
+    `numpy.ndarray`.
 
     - All angular units are now supported, not just `radian`, `degree`
       and `hour`, but now `arcsecond` and `arcminute` as well.  The